REFACTOR-002: extract entry points into ./parse/project.ts; parse.ts becomes a pure barrel
Pure structural move, no behavior change. Moves parseProjectMetadata,
parseSong, parseClip, parseProject, assertProject, isProject,
checkSchemaVersion, and checkProjectIntegrity out of parse.ts into
./parse/project.ts.

With this move parse.ts holds no logic of its own: it only re-exports
the exact same 15 symbols (MAX_TRACK_INSERTS, MAX_RETURN_BUSES,
DomainIssueCode, DomainIssue, ParseResult, DomainValidationError,
formatIssue, parseProjectMetadata, parseSong, parseClip, parseProject,
assertProject, isProject, checkProjectIntegrity, checkSongIntegrity,
checkClipInvariants) it exported before this stack started, so every
one of the 20 existing importers keeps compiling unchanged.

Sixth and final PR of the REFACTOR-002 stack. Closes #143.

Closes #143
